LayoutTests/media/modern-media-controls/overflow-button/overflow-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>OverflowButton</code> class.");

const overflowButton = new OverflowButton({ layoutTraits: new MacOSLayoutTraits(LayoutTraits.Mode.Inline) });

shouldBeEqualToString("overflowButton.element.localName", "button");
shouldBeTrue('overflowButton.element.classList.contains("overflow")');
shouldBe("overflowButton.iconName", "Icons.Overflow");

const iOSOverflowButton = new OverflowButton({ layoutTraits: new IOSLayoutTraits(LayoutTraits.Mode.Inline) });

shouldBecomeEqual('overflowButton.image.element.style.webkitMaskImage.includes("macOS/Overflow.svg")', "true", () => {
    shouldBecomeEqual('iOSOverflowButton.image.element.style.webkitMaskImage.includes("iOS/Overflow.svg")', "true", finishJSTest);
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
